server/requirements.txt
ccxt
pandas
statsmodels
asyncpg
redis
fastapi
uvicorn
python-dotenv
pydantic
scipy
psycopg2-binary
python-socketio
aiohttp
